Add RESTEasy tests
